% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{boom}
\alias{boom}
\title{Print the Output of Intermediate Steps Of a Call}
\usage{
boom(expr)
}
\arguments{
\item{expr}{call to explode}
}
\description{
Print the Output of Intermediate Steps Of a Call
}
\examples{
boom(subset(head(mtcars, 2), qsec > 17))
}
