% Copyright 2011 Google Inc. All Rights Reserved.
% Author: steve.the.bayesian@gmail.com (Steve Scott)

\name{double.model}
\title{Prior distributions for a real valued scalar}
\alias{DoubleModel}
\alias{double.model}

\description{A \code{DoubleModel} is a class of prior distributions for
  real valued scalar parameters.  A \code{DoubleModel} is sometimes used
  by a probability model that does not have a conjugate prior.
}

\seealso{
  \code{\link{BetaPrior}},
  \code{\link{GammaPrior}},
  \code{\link{LognormalPrior}},
  \code{\link{NormalPrior}},
  \code{\link{SdPrior}},
  \code{\link{TruncatedGammaPrior}},
  and \code{\link{UniformPrior}}.
}

\author{
  Steven L. Scott \email{steve.the.bayesian@gmail.com}
}
