% Copyright 2011 Google Inc. All Rights Reserved.
% Author: stevescott@google.com (Steve Scott)

\name{double.model}
\title{Prior distributions for a real valued scalar}
\alias{DoubleModel}
\alias{double.model}

\description{A \code{DoubleModel} is a class of prior distributions for
  real valued scalar parameters.  A \code{DoubleModel} is sometimes used
  by a probability model that does not have a conjugate prior.
}

\seealso{ \code{\link{UniformPrior}}, \code{\link{BetaPrior}},
  \code{\link{NormalPrior}}, \code{\link{LognormalPrior}}, and
  \code{\link{GammaPrior}}.  }

\author{
  Steven L. Scott \email{stevescott@google.com}
}
